Added SQL calls for adding, editing, and deleting sessions.
This commit is a bit late since I did so much. This api basically is every function within the dashboard view. Putting them all in one file is more convienent and fits with the convention set with the login related files. All of these are a combination of checks and basic SQL calls.